Move buffer/store/model/column to buffer/definition/column
