+++ internal/pkg/service/buffer/definition/column/body.go
